git-svn-id: svn://123.57.173.36/meijia-parent@677 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/admin/AdminRoleMapper.java
@@ -10,7 +10,7 @@ public interface AdminRoleMapper {
 
     int insert(AdminRole record);
 
-    int insertSelective(AdminRole record);
+    Long insertSelective(AdminRole record);
 
     AdminRole selectByPrimaryKey(Long id);
